wip(cachecomponents): extract ParkCard schedule footer to a Client Component
The card renders 'closes in X / opens in Y' from the current time; under Cache
Components a server render can't read Date.now(). Move the whole schedule footer
(both open/closed branches) into a 'use client' ParkCardScheduleFooter — client
renders may read the clock, behaviour is identical (suppressHydrationWarning +
client ParkTime were already there). ParkCard stays sync/dual-use.
